+++ face.h
@@ -6,27 +6,6 @@
 #include "trail.h"
 #include "vertex.h"
 
-typedef struct face *FACE;
-
-struct face {
-  // cycle must be null if cycleSetSize is not 1.
-  DYNAMIC struct facialCycle *cycle;
-  DYNAMIC TRAIL backtrack;
-  /* We vertex to previous and next with the same number of colors. */
-  DYNAMIC FACE previous;
-  DYNAMIC FACE next;
-  STATIC COLORSET colors;           // holds up to NFACES
-  DYNAMIC uint_trail cycleSetSize;  // holds up to NCYCLES
-  DYNAMIC CYCLESET_DECLARE possibleCycles;
-  STATIC struct face *adjacentFaces[NCOLORS];
-  STATIC struct edge edges[NCOLORS];
-  /* Precomputed lookup tables for previous and next. */
-  STATIC FACE previousByCycleId[NCYCLES];
-  STATIC FACE nextByCycleId[NCYCLES];
-};
-
-extern STATIC struct face Faces[NFACES];
-
 /* These next two functions are actually defined in search.c */
 extern FAILURE dynamicFaceBacktrackableChoice(FACE face);
 extern FAILURE dynamicFaceChoice(FACE face, int depth);
