+++ initialize.c
@@ -1,7 +1,6 @@
 /* Copyright (C) 2025 Jeremy J. Carroll. See LICENSE for details. */
 
-#include "initialize.h"
-
+#include "dataflow.h"
 #include "face.h"
 #include "graphml.h"
 #include "main.h"
@@ -9,7 +8,6 @@
 #include "s6.h"
 #include "statistics.h"
 #include "utils.h"
-#include "vsearch.h"
 
 #include <stdio.h>
 #include <stdlib.h>
