--- initialize.h
@@ -1,8 +1,0 @@
-/* Copyright (C) 2025 Jeremy J. Carroll. See LICENSE for details. */
-
-#ifndef INITIALIZE_H
-#define INITIALIZE_H
-
-#include "engine.h"
-
-#endif /* INITIALIZE_H */
