+++ save.h
@@ -3,12 +3,11 @@
 #ifndef SAVE_H
 #define SAVE_H
 
-#include "engine.h"
-
 extern int PerFaceDegreeSolutionNumber;
 extern int VariationCount;
 extern int VariationNumber;
 extern int Levels;
 extern char CurrentPrefix[1024];
+extern int searchCountVariations(char *variationMultiplication);
 
 #endif /* SAVE_H */
