+++ test/test_graphml.c
@@ -6,10 +6,10 @@
 #include "main.h"
 #include "predicates.h"
 #include "s6.h"
+#include "save.h"
 #include "statistics.h"
 #include "test_helpers.h"
 #include "utils.h"
-#include "vsearch.h"
 
 #include <regex.h>
 #include <search.h>
