+++ test/test_venn5.c
@@ -6,7 +6,6 @@
 #include "statistics.h"
 #include "test_helpers.h"
 #include "utils.h"
-#include "vsearch.h"
 
 #include <stdio.h>
 #include <stdlib.h>
